Add files via upload
File Name: Data Quality Assessment KPMG.ipynb

Description:
This Jupyter Notebook contains code and documentation for exploratory data analysis (EDA) and data wrangling tasks. It demonstrates the process of analyzing and preprocessing a dataset to gain insights and prepare it for further analysis or modeling.

Contents:
- Loading and cleaning the dataset
- Handling missing values and outliers
- Feature engineering and transformation


Dependencies:
- Python 3.x
- Pandas
- NumPy
- Matplotlib
- Seaborn

Instructions:
1. Install the required dependencies using pip or conda.
2. Download the dataset (provide link or instructions if necessary) and save it in the same directory as this notebook.
3. Run each cell in sequential order to execute the code and generate the results.
4. Modify the code as needed for your own dataset or analysis.



Please note that this file is for educational purposes and serves as an example of EDA and data wrangling techniques. Feel free to adapt and modify the code to suit your own needs.